tags added and many to many relations done but crud not done

+++ app/Http/Controllers/PostController.php
@@ -4,8 +4,11 @@ namespace App\Http\Controllers;
 
 use App\Category;
 use App\Post;
+use App\Tag;
+use App\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Str;
 
 class PostController extends Controller
 {
@@ -17,7 +20,21 @@ class PostController extends Controller
     public function index()
     {
 
-        $posts = Post::orderBy('id', 'desc')->get();
+        $posts = Post::with('category')
+            ->where('user_id', Auth::id())
+            ->orderBy('id', 'desc')
+            ->get();
+//        $posts = Post::with('category:id,name')->orderBy('id', 'desc')->get();
+//        $posts = Post::with(array('category'=>function($query){
+//            $query->select('id', 'name');
+//        }))->orderBy('id', 'desc')->get();
+//        $posts = Post::orderBy('id', 'desc')->get();
+        //dd($posts);
+//        $user = auth()->user();
+//        $user->load('posts');
+//        $posts =$user->posts;
+//        $posts = auth()->user()->posts;
+
         return view('admin.post.index', compact('posts'));
     }
 
@@ -29,7 +46,8 @@ class PostController extends Controller
     public function create()
     {
         $categories = Category::all();
-        return view('admin.post.create', compact('categories'));
+        $tags = Tag::all();
+        return view('admin.post.create', compact('categories', 'tags'));
     }
 
     /**
@@ -40,21 +58,46 @@ class PostController extends Controller
      */
     public function store(Request $request)
     {
-
+//        dd($request->all());
         //validation
-        $this->validate($request, ['title' => 'required']);
+        $this->validate($request, [
+            'title' => 'required',
+//            'img' => 'required|mimes:png|max:300'
+        ]);
 
         $data = $request->all();
         $data['user_id'] = Auth::id();
         if ($request->has('img')) {
             $file = $request->file('img');
-            $extension = $file->getClientOriginalExtension(); //getting file extension
+
+            $extension = $file->getClientOriginalExtension();//getting file extension
+
             $filename = time() . '.' . $extension;
             $file->move('uploads/', $filename);
             $data['img'] = $filename;
+
+
+//            //second method to upload image
+//            $filename = Str::slug($data['title'])  . '-'. time() .  '.' . $file->getClientOriginalExtension();
+//
+////            dd($data);
+//            $file->storePubliclyAs(
+//                'public/posts',  $filename
+//            );
+//
+//            $data['img'] = $filename;
         }
+
+
         // dd($data);
-        Post::create($data);
+        $post = Post::create($data);
+
+        //multiple tags insert tag_id array
+        if(!empty($request->tag_id)){
+            $post->tags()->sync($request->tag_id);
+        }
+
+
         return redirect('/posts')->with('status', "Created Successfully");
     }
 
@@ -80,7 +123,8 @@ class PostController extends Controller
 
     {
         $categories = Category::all();
-        return view('admin.post.edit', compact('post', 'categories'));
+        $tags = Tag::all();
+        return view('admin.post.edit', compact('post', 'categories','tags'));
 
     }
 
@@ -98,23 +142,30 @@ class PostController extends Controller
         $data = $request->all();
 
 
-
-
         if ($request->has('img')) {
 
             $file_path = public_path('/uploads/' . $post->img);
-
-            unlink($file_path);
+            //dd($post->img);
+            if (file_exists($file_path) && !empty($post->img)) {
+                unlink($file_path);
+            }
 
             $file = $request->file('img');
             $extension = $file->getClientOriginalExtension(); //getting file extension
             $filename = time() . '.' . $extension;
             $file->move('uploads/', $filename);
             $data['img'] = $filename;
+
         }
 
 
         $post->update($data);
+
+        //multiple tags insert
+        if(!empty($request->tag_id)){
+            $post->tags()->sync($request->tag_id);
+        }
+
         return redirect('/posts')->with('status', "Updated Successfully");
     }
